requirements.txt
aiofiles
aiohttp
asyncio
bing-image-urls # this for in image module for searching images from bing
dnspython==2.3.0
ffmpeg-python
gitpython
gTTS
googletrans
googlesearch-python
google_play_scraper #for playstore app search 
hachoir
httpx[http2]
heroku3
img2pdf
lyricsgenius
lexica-api
MukeshAPI
motor
pillow==9.5.0
pyfiglet
opencv-python-headless==4.9.0.80
profanity
pytz
psutil
pytube
pykeyboard
py-tgcalls==2.0.1
#pyrogram
pyrofork
python-dotenv
pyyaml
requests
speedtest-cli
secureme==0.0.4
spotipy
telethon==1.33.1
telegraph==2.2.0
tgcrypto
urllib3
yt-dlp
youtube-search-python
#youtube_search
#git+https://github.com/alexmercerind/youtube-search-python@main
git+https://github.com/joetats/youtube_search@master
uuid==1.30
SafoneAPI